# Bump version of protozero to 1.7.1