… third_party/protozero/.clang-tidy
… third_party/protozero/.github/actions/build-windows/action.yml
… third_party/protozero/.github/actions/build/action.yml
… third_party/protozero/.github/actions/cmake-windows/action.yml
… third_party/protozero/.github/actions/cmake/action.yml
… third_party/protozero/.github/actions/ctest-windows/action.yml
… third_party/protozero/.github/actions/ctest/action.yml
… third_party/protozero/.github/actions/install-ubuntu/action.yml
… third_party/protozero/.github/workflows/ci.yml
… third_party/protozero/CHANGELOG.md
… third_party/protozero/CMakeLists.txt
… third_party/protozero/FUZZING.md
… third_party/protozero/README.md
… third_party/protozero/appveyor.yml
… third_party/protozero/bench/data/README.md
… third_party/protozero/doc/Doxyfile.in
… third_party/protozero/include/protozero/buffer_string.hpp
… third_party/protozero/include/protozero/buffer_vector.hpp
… third_party/protozero/include/protozero/byteswap.hpp
… third_party/protozero/include/protozero/version.hpp
… third_party/protozero/test/CMakeLists.txt
… third_party/protozero/test/README.md
… third_party/protozero/test/catch/catch.hpp
… third_party/protozero/test/create_pbf_test_data.sh
… third_party/protozero/test/include/buffer.hpp
… third_party/protozero/test/include/packed_access.hpp
… third_party/protozero/test/include/scalar_access.hpp
… third_party/protozero/test/include/test.hpp
… third_party/protozero/test/include/testcase.hpp
… third_party/protozero/test/reader_tests.cpp