<!-- BENCHMARK_RESULTS_START -->
<details><summary><h2>Benchmark Results</h2></summary>

| Benchmark | Base | PR |
|-----------|------|----|
| alias | aliased u32: 1092.55<br/>plain u32: 1097.59<br/>aliased double: 952.053<br/>plain double: 965.074 | aliased u32: 1094.14<br/>plain u32: 1091.07<br/>aliased double: 959.891<br/>plain double: 962.612 |
| e2e_match_ch | Ops: 41.86 ± 0.07 ops/s. Best: 41.97 ops/s<br/>Total: 3129.69ms ± 6.22ms. Best: 3121.41ms<br/>Min time: 2.18ms ± 0.03ms<br/>Mean time: 23.89ms ± 0.05ms<br/>Median time: 18.01ms ± 0.16ms<br/>95th percentile: 80.07ms ± 0.38ms<br/>99th percentile: 98.48ms ± 0.60ms<br/>Max time: 105.24ms ± 0.42ms | Ops: 41.40 ± 0.06 ops/s. Best: 41.48 ops/s<br/>Total: 3164.38ms ± 5.00ms. Best: 3158.10ms<br/>Min time: 2.16ms ± 0.02ms<br/>Mean time: 24.16ms ± 0.04ms<br/>Median time: 17.77ms ± 0.08ms<br/>95th percentile: 83.19ms ± 0.33ms<br/>99th percentile: 102.48ms ± 0.43ms<br/>Max time: 107.07ms ± 0.49ms |
| e2e_match_mld | Ops: 64.22 ± 0.11 ops/s. Best: 64.36 ops/s<br/>Total: 2039.89ms ± 3.64ms. Best: 2035.38ms<br/>Min time: 1.73ms ± 0.01ms<br/>Mean time: 15.57ms ± 0.03ms<br/>Median time: 8.06ms ± 0.06ms<br/>95th percentile: 52.15ms ± 0.20ms<br/>99th percentile: 60.52ms ± 0.19ms<br/>Max time: 69.14ms ± 0.09ms | Ops: 63.99 ± 0.13 ops/s. Best: 64.16 ops/s<br/>Total: 2047.13ms ± 4.28ms. Best: 2041.82ms<br/>Min time: 1.75ms ± 0.02ms<br/>Mean time: 15.63ms ± 0.03ms<br/>Median time: 8.11ms ± 0.11ms<br/>95th percentile: 52.45ms ± 0.27ms<br/>99th percentile: 60.95ms ± 0.49ms<br/>Max time: 69.39ms ± 0.38ms |
| e2e_nearest_ch | Ops: 877.68 ± 7.18 ops/s. Best: 886.82 ops/s<br/>Total: 1139.62ms ± 9.57ms. Best: 1127.63ms<br/>Min time: 0.97ms ± 0.00ms<br/>Mean time: 1.14ms ± 0.01ms<br/>Median time: 1.05ms ± 0.00ms<br/>95th percentile: 1.58ms ± 0.01ms<br/>99th percentile: 1.63ms ± 0.02ms<br/>Max time: 5.73ms ± 2.67ms | Ops: 874.82 ± 4.63 ops/s. Best: 883.57 ops/s<br/>Total: 1143.13ms ± 6.16ms. Best: 1131.77ms<br/>Min time: 0.98ms ± 0.00ms<br/>Mean time: 1.14ms ± 0.01ms<br/>Median time: 1.06ms ± 0.00ms<br/>95th percentile: 1.59ms ± 0.01ms<br/>99th percentile: 1.64ms ± 0.01ms<br/>Max time: 6.04ms ± 2.95ms |
| e2e_nearest_mld | Ops: 878.16 ± 2.59 ops/s. Best: 884.22 ops/s<br/>Total: 1138.68ms ± 3.46ms. Best: 1130.94ms<br/>Min time: 0.96ms ± 0.01ms<br/>Mean time: 1.14ms ± 0.00ms<br/>Median time: 1.05ms ± 0.00ms<br/>95th percentile: 1.58ms ± 0.00ms<br/>99th percentile: 1.63ms ± 0.01ms<br/>Max time: 5.80ms ± 2.69ms | Ops: 878.13 ± 2.16 ops/s. Best: 881.74 ops/s<br/>Total: 1138.79ms ± 2.91ms. Best: 1134.12ms<br/>Min time: 0.97ms ± 0.01ms<br/>Mean time: 1.14ms ± 0.00ms<br/>Median time: 1.05ms ± 0.00ms<br/>95th percentile: 1.58ms ± 0.01ms<br/>99th percentile: 1.63ms ± 0.00ms<br/>Max time: 6.08ms ± 2.86ms |
| e2e_route_ch | Ops: 367.73 ± 1.06 ops/s. Best: 369.02 ops/s<br/>Total: 2719.63ms ± 8.08ms. Best: 2709.91ms<br/>Min time: 1.18ms ± 0.01ms<br/>Mean time: 2.72ms ± 0.01ms<br/>Median time: 2.73ms ± 0.01ms<br/>95th percentile: 3.60ms ± 0.01ms<br/>99th percentile: 4.01ms ± 0.07ms<br/>Max time: 6.90ms ± 2.49ms | Ops: 369.91 ± 1.24 ops/s. Best: 371.45 ops/s<br/>Total: 2703.07ms ± 9.18ms. Best: 2692.12ms<br/>Min time: 1.17ms ± 0.01ms<br/>Mean time: 2.70ms ± 0.01ms<br/>Median time: 2.72ms ± 0.01ms<br/>95th percentile: 3.58ms ± 0.01ms<br/>99th percentile: 3.96ms ± 0.02ms<br/>Max time: 6.54ms ± 2.15ms |
| e2e_route_mld | Ops: 305.55 ± 0.61 ops/s. Best: 306.46 ops/s<br/>Total: 3272.54ms ± 6.02ms. Best: 3263.04ms<br/>Min time: 1.16ms ± 0.01ms<br/>Mean time: 3.27ms ± 0.01ms<br/>Median time: 3.32ms ± 0.01ms<br/>95th percentile: 4.50ms ± 0.01ms<br/>99th percentile: 4.98ms ± 0.06ms<br/>Max time: 7.37ms ± 1.95ms | Ops: 305.26 ± 0.53 ops/s. Best: 305.93 ops/s<br/>Total: 3276.10ms ± 5.81ms. Best: 3268.67ms<br/>Min time: 1.16ms ± 0.02ms<br/>Mean time: 3.28ms ± 0.01ms<br/>Median time: 3.31ms ± 0.01ms<br/>95th percentile: 4.50ms ± 0.03ms<br/>99th percentile: 5.02ms ± 0.05ms<br/>Max time: 7.51ms ± 1.97ms |
| e2e_table_ch | Ops: 317.37 ± 1.06 ops/s. Best: 318.55 ops/s<br/>Total: 3150.55ms ± 10.67ms. Best: 3139.20ms<br/>Min time: 1.62ms ± 0.02ms<br/>Mean time: 3.15ms ± 0.01ms<br/>Median time: 3.14ms ± 0.01ms<br/>95th percentile: 4.37ms ± 0.02ms<br/>99th percentile: 4.71ms ± 0.06ms<br/>Max time: 8.92ms ± 2.89ms | Ops: 321.63 ± 0.50 ops/s. Best: 322.37 ops/s<br/>Total: 3109.23ms ± 5.02ms. Best: 3102.04ms<br/>Min time: 1.56ms ± 0.03ms<br/>Mean time: 3.11ms ± 0.01ms<br/>Median time: 3.10ms ± 0.01ms<br/>95th percentile: 4.33ms ± 0.02ms<br/>99th percentile: 4.66ms ± 0.06ms<br/>Max time: 8.18ms ± 2.46ms |
| e2e_table_mld | Ops: 109.89 ± 0.16 ops/s. Best: 110.07 ops/s<br/>Total: 9099.99ms ± 12.46ms. Best: 9085.40ms<br/>Min time: 3.65ms ± 0.07ms<br/>Mean time: 9.10ms ± 0.01ms<br/>Median time: 9.06ms ± 0.05ms<br/>95th percentile: 13.94ms ± 0.01ms<br/>99th percentile: 14.66ms ± 0.05ms<br/>Max time: 17.34ms ± 2.52ms | Ops: 110.43 ± 0.13 ops/s. Best: 110.65 ops/s<br/>Total: 9055.84ms ± 10.85ms. Best: 9037.10ms<br/>Min time: 3.68ms ± 0.07ms<br/>Mean time: 9.06ms ± 0.01ms<br/>Median time: 9.03ms ± 0.03ms<br/>95th percentile: 13.86ms ± 0.02ms<br/>99th percentile: 14.64ms ± 0.05ms<br/>Max time: 17.36ms ± 2.37ms |
| e2e_trip_ch | Ops: 99.24 ± 0.34 ops/s. Best: 99.68 ops/s<br/>Total: 10075.99ms ± 35.23ms. Best: 10032.61ms<br/>Min time: 1.44ms ± 0.13ms<br/>Mean time: 10.08ms ± 0.04ms<br/>Median time: 9.53ms ± 0.05ms<br/>95th percentile: 18.20ms ± 0.08ms<br/>99th percentile: 20.12ms ± 0.09ms<br/>Max time: 22.29ms ± 0.69ms | Ops: 99.65 ± 0.39 ops/s. Best: 100.04 ops/s<br/>Total: 10035.64ms ± 42.69ms. Best: 9996.46ms<br/>Min time: 1.44ms ± 0.10ms<br/>Mean time: 10.04ms ± 0.04ms<br/>Median time: 9.48ms ± 0.07ms<br/>95th percentile: 18.16ms ± 0.06ms<br/>99th percentile: 20.06ms ± 0.12ms<br/>Max time: 21.84ms ± 0.55ms |
| e2e_trip_mld | Ops: 58.67 ± 0.10 ops/s. Best: 58.89 ops/s<br/>Total: 17043.87ms ± 27.79ms. Best: 16979.44ms<br/>Min time: 1.65ms ± 0.27ms<br/>Mean time: 17.04ms ± 0.03ms<br/>Median time: 16.56ms ± 0.06ms<br/>95th percentile: 28.06ms ± 0.06ms<br/>99th percentile: 30.32ms ± 0.14ms<br/>Max time: 31.83ms ± 0.07ms | Ops: 58.39 ± 0.20 ops/s. Best: 58.70 ops/s<br/>Total: 17126.33ms ± 60.20ms. Best: 17035.19ms<br/>Min time: 1.63ms ± 0.23ms<br/>Mean time: 17.13ms ± 0.07ms<br/>Median time: 16.56ms ± 0.09ms<br/>95th percentile: 28.15ms ± 0.08ms<br/>99th percentile: 30.44ms ± 0.19ms<br/>Max time: 37.38ms ± 8.11ms |
| json-render | String: 5.77834ms<br/>Stringstream: 9.22759ms<br/>Vector: 6.56206ms | String: 5.57293ms<br/>Stringstream: 9.01642ms<br/>Vector: 6.58281ms |
| match_ch | Default radius: <br/>4.62467ms/req at 82 coordinate<br/>0.0563984ms/coordinate<br/>Radius 10m: <br/>16.1721ms/req at 82 coordinate<br/>0.19722ms/coordinate | Default radius: <br/>4.62884ms/req at 82 coordinate<br/>0.0564492ms/coordinate<br/>Radius 10m: <br/>16.2961ms/req at 82 coordinate<br/>0.198733ms/coordinate |
| match_mld | Default radius: <br/>3.15431ms/req at 82 coordinate<br/>0.0384672ms/coordinate<br/>Radius 10m: <br/>10.9579ms/req at 82 coordinate<br/>0.133633ms/coordinate | Default radius: <br/>3.06678ms/req at 82 coordinate<br/>0.0373997ms/coordinate<br/>Radius 10m: <br/>11.123ms/req at 82 coordinate<br/>0.135646ms/coordinate |
| osrm_contract | Time: 98.89s Peak RAM: 200.97MB | Time: 99.74s Peak RAM: 200.29MB |
| osrm_customize | Time: 1.32s Peak RAM: 117.58MB | Time: 1.30s Peak RAM: 117.69MB |
| osrm_extract | Time: 11.84s Peak RAM: 428.12MB | Time: 11.72s Peak RAM: 427.08MB |
| osrm_partition | Time: 2.11s Peak RAM: 144.87MB | Time: 2.11s Peak RAM: 144.87MB |
| packedvector | random write:<br/>std::vector 9848.8 ms<br/>util::packed_vector 73640.8 ms<br/>slowdown: 7.47713<br/>random read:<br/>std::vector 8515.36 ms<br/>util::packed_vector 30339.3 ms<br/>slowdown: 3.56289 | random write:<br/>std::vector 9856.87 ms<br/>util::packed_vector 73333.5 ms<br/>slowdown: 7.43984<br/>random read:<br/>std::vector 8497.53 ms<br/>util::packed_vector 30137.6 ms<br/>slowdown: 3.54663 |
| random_match_ch | 500 matches, default radius<br/>ops: 203.68 ± 0.68 ops/s. best: 204.40ops/s.<br/>total: 279.86 ± 0.93ms. best: 278.87ms.<br/>avg: 4.91 ± 0.02ms<br/>min: 0.13 ± 0.01ms<br/>max: 24.43 ± 0.18ms<br/>p99: 24.43 ± 0.18ms<br/><br/>500 matches, radius=10<br/>ops: 59.58 ± 0.07 ops/s. best: 59.69ops/s.<br/>total: 1074.23 ± 1.22ms. best: 1072.25ms.<br/>avg: 16.78 ± 0.02ms<br/>min: 0.15 ± 0.00ms<br/>max: 228.04 ± 0.80ms<br/>p99: 228.04 ± 0.80ms<br/><br/>500 matches, radius=20<br/>ops: 14.30 ± 0.03 ops/s. best: 14.34ops/s.<br/>total: 4544.20 ± 8.53ms. best: 4531.63ms.<br/>avg: 69.91 ± 0.13ms<br/>min: 0.31 ± 0.01ms<br/>max: 1180.74 ± 4.80ms<br/>p99: 1180.74 ± 4.80ms | 500 matches, default radius<br/>ops: 204.17 ± 0.75 ops/s. best: 205.00ops/s.<br/>total: 279.19 ± 1.07ms. best: 278.04ms.<br/>avg: 4.90 ± 0.02ms<br/>min: 0.14 ± 0.01ms<br/>max: 24.72 ± 0.11ms<br/>p99: 24.72 ± 0.11ms<br/><br/>500 matches, radius=10<br/>ops: 60.05 ± 0.17 ops/s. best: 60.31ops/s.<br/>total: 1065.81 ± 2.88ms. best: 1061.15ms.<br/>avg: 16.65 ± 0.05ms<br/>min: 0.15 ± 0.00ms<br/>max: 222.91 ± 1.31ms<br/>p99: 222.91 ± 1.31ms<br/><br/>500 matches, radius=20<br/>ops: 14.41 ± 0.01 ops/s. best: 14.43ops/s.<br/>total: 4509.92 ± 3.70ms. best: 4504.26ms.<br/>avg: 69.38 ± 0.06ms<br/>min: 0.30 ± 0.00ms<br/>max: 1138.46 ± 4.12ms<br/>p99: 1138.46 ± 4.12ms |
| random_match_mld | 500 matches, default radius<br/>ops: 301.79 ± 2.30 ops/s. best: 304.20ops/s.<br/>total: 188.89 ± 1.44ms. best: 187.38ms.<br/>avg: 3.31 ± 0.03ms<br/>min: 0.12 ± 0.01ms<br/>max: 19.01 ± 0.17ms<br/>p99: 19.01 ± 0.17ms<br/><br/>500 matches, radius=10<br/>ops: 106.28 ± 0.18 ops/s. best: 106.48ops/s.<br/>total: 602.17 ± 1.03ms. best: 601.08ms.<br/>avg: 9.41 ± 0.02ms<br/>min: 0.14 ± 0.00ms<br/>max: 111.17 ± 0.38ms<br/>p99: 111.17 ± 0.38ms<br/><br/>500 matches, radius=20<br/>ops: 21.63 ± 0.02 ops/s. best: 21.67ops/s.<br/>total: 3004.80 ± 2.65ms. best: 2999.50ms.<br/>avg: 46.23 ± 0.04ms<br/>min: 0.19 ± 0.00ms<br/>max: 584.85 ± 1.27ms<br/>p99: 584.85 ± 1.27ms | 500 matches, default radius<br/>ops: 302.47 ± 1.76 ops/s. best: 304.83ops/s.<br/>total: 188.46 ± 1.21ms. best: 186.99ms.<br/>avg: 3.31 ± 0.02ms<br/>min: 0.12 ± 0.01ms<br/>max: 18.96 ± 0.05ms<br/>p99: 18.96 ± 0.05ms<br/><br/>500 matches, radius=10<br/>ops: 106.94 ± 0.12 ops/s. best: 107.16ops/s.<br/>total: 598.49 ± 0.65ms. best: 597.23ms.<br/>avg: 9.35 ± 0.01ms<br/>min: 0.14 ± 0.00ms<br/>max: 110.66 ± 0.25ms<br/>p99: 110.66 ± 0.25ms<br/><br/>500 matches, radius=20<br/>ops: 21.66 ± 0.02 ops/s. best: 21.69ops/s.<br/>total: 3001.04 ± 2.31ms. best: 2996.36ms.<br/>avg: 46.17 ± 0.04ms<br/>min: 0.19 ± 0.00ms<br/>max: 584.21 ± 0.81ms<br/>p99: 584.21 ± 0.81ms |
| random_nearest_ch | 10000 nearest, number_of_results=1<br/>ops: 24612.69 ± 69.12 ops/s. best: 24682.55ops/s.<br/>total: 406.30 ± 1.14ms. best: 405.14ms.<br/>avg: 0.04 ± 0.00ms<br/>min: 0.01 ± 0.00ms<br/>max: 0.16 ± 0.03ms<br/>p99: 0.10 ± 0.00ms<br/><br/>10000 nearest, number_of_results=5<br/>ops: 18452.36 ± 38.65 ops/s. best: 18508.20ops/s.<br/>total: 541.94 ± 1.14ms. best: 540.30ms.<br/>avg: 0.05 ± 0.00ms<br/>min: 0.02 ± 0.00ms<br/>max: 0.15 ± 0.00ms<br/>p99: 0.11 ± 0.00ms<br/><br/>10000 nearest, number_of_results=10<br/>ops: 14644.65 ± 22.95 ops/s. best: 14668.39ops/s.<br/>total: 682.85 ± 1.11ms. best: 681.74ms.<br/>avg: 0.07 ± 0.00ms<br/>min: 0.03 ± 0.00ms<br/>max: 0.17 ± 0.01ms<br/>p99: 0.13 ± 0.00ms | 10000 nearest, number_of_results=1<br/>ops: 24680.09 ± 98.30 ops/s. best: 24784.21ops/s.<br/>total: 405.19 ± 1.62ms. best: 403.48ms.<br/>avg: 0.04 ± 0.00ms<br/>min: 0.01 ± 0.00ms<br/>max: 0.15 ± 0.03ms<br/>p99: 0.10 ± 0.00ms<br/><br/>10000 nearest, number_of_results=5<br/>ops: 18546.31 ± 26.26 ops/s. best: 18584.19ops/s.<br/>total: 539.19 ± 0.76ms. best: 538.09ms.<br/>avg: 0.05 ± 0.00ms<br/>min: 0.02 ± 0.00ms<br/>max: 0.22 ± 0.11ms<br/>p99: 0.11 ± 0.00ms<br/><br/>10000 nearest, number_of_results=10<br/>ops: 14784.71 ± 27.23 ops/s. best: 14825.73ops/s.<br/>total: 676.38 ± 1.25ms. best: 674.50ms.<br/>avg: 0.07 ± 0.00ms<br/>min: 0.03 ± 0.00ms<br/>max: 0.17 ± 0.00ms<br/>p99: 0.13 ± 0.00ms |
| random_nearest_mld | 10000 nearest, number_of_results=1<br/>ops: 24690.34 ± 79.20 ops/s. best: 24800.71ops/s.<br/>total: 405.02 ± 1.41ms. best: 403.21ms.<br/>avg: 0.04 ± 0.00ms<br/>min: 0.01 ± 0.00ms<br/>max: 0.16 ± 0.02ms<br/>p99: 0.10 ± 0.00ms<br/><br/>10000 nearest, number_of_results=5<br/>ops: 18546.77 ± 25.27 ops/s. best: 18578.74ops/s.<br/>total: 539.18 ± 0.77ms. best: 538.25ms.<br/>avg: 0.05 ± 0.00ms<br/>min: 0.02 ± 0.00ms<br/>max: 0.15 ± 0.00ms<br/>p99: 0.11 ± 0.00ms<br/><br/>10000 nearest, number_of_results=10<br/>ops: 14652.02 ± 26.57 ops/s. best: 14692.80ops/s.<br/>total: 682.50 ± 1.24ms. best: 680.61ms.<br/>avg: 0.07 ± 0.00ms<br/>min: 0.03 ± 0.00ms<br/>max: 0.18 ± 0.01ms<br/>p99: 0.13 ± 0.00ms | 10000 nearest, number_of_results=1<br/>ops: 24499.70 ± 52.06 ops/s. best: 24564.53ops/s.<br/>total: 408.17 ± 0.87ms. best: 407.09ms.<br/>avg: 0.04 ± 0.00ms<br/>min: 0.01 ± 0.00ms<br/>max: 0.15 ± 0.02ms<br/>p99: 0.10 ± 0.00ms<br/><br/>10000 nearest, number_of_results=5<br/>ops: 18454.30 ± 7.56 ops/s. best: 18468.92ops/s.<br/>total: 541.88 ± 0.22ms. best: 541.45ms.<br/>avg: 0.05 ± 0.00ms<br/>min: 0.02 ± 0.00ms<br/>max: 0.15 ± 0.00ms<br/>p99: 0.11 ± 0.00ms<br/><br/>10000 nearest, number_of_results=10<br/>ops: 14697.83 ± 15.18 ops/s. best: 14718.19ops/s.<br/>total: 680.37 ± 0.72ms. best: 679.43ms.<br/>avg: 0.07 ± 0.00ms<br/>min: 0.03 ± 0.00ms<br/>max: 0.17 ± 0.01ms<br/>p99: 0.13 ± 0.00ms |
| random_route_ch | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>ops: 509.44 ± 2.31 ops/s. best: 512.71ops/s.<br/>total: 1931.58 ± 8.77ms. best: 1919.21ms.<br/>avg: 1.96 ± 0.01ms<br/>min: 0.32 ± 0.01ms<br/>max: 3.48 ± 0.29ms<br/>p99: 2.84 ± 0.06ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>ops: 572.84 ± 8.33 ops/s. best: 579.92ops/s.<br/>total: 1746.30 ± 26.16ms. best: 1724.37ms.<br/>avg: 1.75 ± 0.03ms<br/>min: 0.06 ± 0.00ms<br/>max: 5.55 ± 0.51ms<br/>p99: 4.22 ± 0.10ms<br/><br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>ops: 984.60 ± 12.92 ops/s. best: 1003.60ops/s.<br/>total: 999.61 ± 13.02ms. best: 980.47ms.<br/>avg: 1.02 ± 0.01ms<br/>min: 0.24 ± 0.01ms<br/>max: 1.73 ± 0.07ms<br/>p99: 1.42 ± 0.04ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>ops: 1037.91 ± 4.75 ops/s. best: 1045.91ops/s.<br/>total: 963.50 ± 4.40ms. best: 956.11ms.<br/>avg: 0.96 ± 0.00ms<br/>min: 0.04 ± 0.00ms<br/>max: 2.79 ± 0.09ms<br/>p99: 2.52 ± 0.01ms | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>ops: 517.45 ± 1.40 ops/s. best: 519.89ops/s.<br/>total: 1901.64 ± 5.14ms. best: 1892.70ms.<br/>avg: 1.93 ± 0.01ms<br/>min: 0.32 ± 0.01ms<br/>max: 3.35 ± 0.34ms<br/>p99: 2.75 ± 0.03ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>ops: 628.40 ± 2.11 ops/s. best: 632.08ops/s.<br/>total: 1591.37 ± 5.34ms. best: 1582.07ms.<br/>avg: 1.59 ± 0.01ms<br/>min: 0.06 ± 0.01ms<br/>max: 4.03 ± 0.05ms<br/>p99: 3.30 ± 0.04ms<br/><br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>ops: 1012.04 ± 4.80 ops/s. best: 1019.18ops/s.<br/>total: 972.33 ± 4.61ms. best: 965.48ms.<br/>avg: 0.99 ± 0.00ms<br/>min: 0.23 ± 0.00ms<br/>max: 1.56 ± 0.04ms<br/>p99: 1.36 ± 0.01ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>ops: 1200.76 ± 8.38 ops/s. best: 1208.81ops/s.<br/>total: 832.86 ± 5.85ms. best: 827.26ms.<br/>avg: 0.83 ± 0.01ms<br/>min: 0.04 ± 0.00ms<br/>max: 2.26 ± 0.20ms<br/>p99: 1.88 ± 0.02ms |
| random_route_mld | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>ops: 249.02 ± 1.14 ops/s. best: 250.67ops/s.<br/>total: 3951.60 ± 18.14ms. best: 3925.41ms.<br/>avg: 4.02 ± 0.02ms<br/>min: 0.32 ± 0.00ms<br/>max: 9.00 ± 0.20ms<br/>p99: 6.84 ± 0.04ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>ops: 238.32 ± 0.76 ops/s. best: 239.58ops/s.<br/>total: 4196.10 ± 13.40ms. best: 4173.94ms.<br/>avg: 4.20 ± 0.01ms<br/>min: 0.05 ± 0.00ms<br/>max: 10.30 ± 0.39ms<br/>p99: 8.66 ± 0.04ms<br/><br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>ops: 327.68 ± 1.40 ops/s. best: 329.35ops/s.<br/>total: 3003.02 ± 12.86ms. best: 2987.68ms.<br/>avg: 3.05 ± 0.01ms<br/>min: 0.27 ± 0.00ms<br/>max: 7.30 ± 0.09ms<br/>p99: 5.23 ± 0.05ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>ops: 295.19 ± 0.74 ops/s. best: 296.46ops/s.<br/>total: 3387.72 ± 7.94ms. best: 3373.11ms.<br/>avg: 3.39 ± 0.01ms<br/>min: 0.04 ± 0.00ms<br/>max: 7.55 ± 0.32ms<br/>p99: 6.78 ± 0.05ms | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>ops: 250.06 ± 0.84 ops/s. best: 251.45ops/s.<br/>total: 3935.11 ± 13.27ms. best: 3913.35ms.<br/>avg: 4.00 ± 0.01ms<br/>min: 0.32 ± 0.01ms<br/>max: 8.81 ± 0.15ms<br/>p99: 6.79 ± 0.05ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>ops: 239.08 ± 0.41 ops/s. best: 239.74ops/s.<br/>total: 4182.72 ± 7.01ms. best: 4171.19ms.<br/>avg: 4.18 ± 0.01ms<br/>min: 0.05 ± 0.00ms<br/>max: 9.84 ± 0.23ms<br/>p99: 8.60 ± 0.06ms<br/><br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>ops: 330.23 ± 0.96 ops/s. best: 331.77ops/s.<br/>total: 2979.78 ± 8.73ms. best: 2965.94ms.<br/>avg: 3.03 ± 0.01ms<br/>min: 0.26 ± 0.00ms<br/>max: 7.25 ± 0.05ms<br/>p99: 5.24 ± 0.02ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>ops: 296.19 ± 1.20 ops/s. best: 298.59ops/s.<br/>total: 3376.24 ± 13.63ms. best: 3349.04ms.<br/>avg: 3.38 ± 0.01ms<br/>min: 0.04 ± 0.00ms<br/>max: 7.70 ± 0.27ms<br/>p99: 6.78 ± 0.02ms |
| random_table_ch | 250 tables, 3 coordinates<br/>ops: 1425.02 ± 14.28 ops/s. best: 1437.63ops/s.<br/>total: 175.46 ± 1.78ms. best: 173.90ms.<br/>avg: 0.70 ± 0.01ms<br/>min: 0.53 ± 0.00ms<br/>max: 1.06 ± 0.27ms<br/>p99: 0.88 ± 0.05ms<br/><br/>250 tables, 25 coordinates<br/>ops: 166.98 ± 0.16 ops/s. best: 167.22ops/s.<br/>total: 1497.16 ± 1.42ms. best: 1495.07ms.<br/>avg: 5.99 ± 0.01ms<br/>min: 5.44 ± 0.01ms<br/>max: 6.52 ± 0.02ms<br/>p99: 6.47 ± 0.02ms<br/><br/>250 tables, 50 coordinates<br/>ops: 82.55 ± 0.17 ops/s. best: 82.80ops/s.<br/>total: 3028.56 ± 6.20ms. best: 3019.26ms.<br/>avg: 12.11 ± 0.02ms<br/>min: 11.37 ± 0.03ms<br/>max: 13.20 ± 0.12ms<br/>p99: 12.94 ± 0.06ms | 250 tables, 3 coordinates<br/>ops: 1445.65 ± 12.33 ops/s. best: 1455.91ops/s.<br/>total: 172.95 ± 1.50ms. best: 171.71ms.<br/>avg: 0.69 ± 0.01ms<br/>min: 0.46 ± 0.00ms<br/>max: 1.07 ± 0.25ms<br/>p99: 0.87 ± 0.02ms<br/><br/>250 tables, 25 coordinates<br/>ops: 169.26 ± 0.13 ops/s. best: 169.49ops/s.<br/>total: 1476.99 ± 1.11ms. best: 1475.01ms.<br/>avg: 5.91 ± 0.00ms<br/>min: 5.18 ± 0.01ms<br/>max: 6.62 ± 0.20ms<br/>p99: 6.41 ± 0.01ms<br/><br/>250 tables, 50 coordinates<br/>ops: 83.72 ± 0.06 ops/s. best: 83.82ops/s.<br/>total: 2986.14 ± 2.03ms. best: 2982.54ms.<br/>avg: 11.94 ± 0.01ms<br/>min: 11.03 ± 0.02ms<br/>max: 12.98 ± 0.03ms<br/>p99: 12.83 ± 0.05ms |
…