@SiarheiFedartsou 2025-09-13
@afarber most likely it was intentionally, but honestly I don't remember details...