@afarber 2025-09-13
On the other side the protozero/test folder is only 2 MB and keeping it might make the next third_party/protozero updates simpler, because less differences